création de FormValidation Trait pour valider les données côté serveur; backup profil sans réservation, fix des données manquantes avec création rental lors de la redirection vers le profile/rentals

+++ src/Views/profile/rentals/index.php
@@ -8,27 +8,33 @@
 </head>
 
 <body>
+  <h1>Page user rentals</h1>
   <?php
-  // var_dump($rentals);
+
 
   if (isset($message) && !empty($message)) {
     echo "<p> $message </p>";
   }
 
-  foreach ($rentals as $rental) {
-    echo "<div>";
-    foreach ($rental as $key => $value) {
-      # code...
-      echo "<p> $key : $value </p>";
+
+  if (count($rentals) == 0) {
+    echo "<h2> Vous n'avez pas encore loué de véhicule</h2>";
+    echo "<a href='/location/public/vehicules/index.php'>Voir les véhicules</a>";
+  } else {
+    foreach ($rentals as $rental) {
+      echo "<div>";
+      foreach ($rental as $key => $value) {
+        # code...
+        echo "<p> $key : $value </p>";
+      }
+      $rental_id = $rental['reservation_id'];
+      echo "<a href='/location/public/profile/rentals/update/$rental_id'>Update reservation</a>";
+      echo "<a href='/location/public/profile/rentals/delete/$rental_id'>Delete reservation</a>";
+      echo "</div>";
     }
-    $rental_id = $rental['reservation_id'];
-    echo "<a href='/location/public/profile/rentals/update/$rental_id'>Update reservation</a>";
-    echo "<a href='/location/public/profile/rentals/delete/$rental_id'>Delete reservation</a>";
-    echo "</div>";
   }
 
   ?>
-  <h1>Page user rentals</h1>
 </body>
 
 </html>
